更新依赖版本至0.12.0，重构转录工作者以使用共享的转录引擎，优化音频转录逻辑，增强错误处理，确保更高效的任务处理和模型管理。

+++ voice-cli/src/services/model_service.rs
@@ -293,9 +293,9 @@ impl ModelService {
         ) {
             let actual_size = metadata.len();
             let size_diff_percent = if actual_size > expected_size {
-                ((actual_size - expected_size) as f64 / expected_size as f64) * 100.0
+                ((actual_size as f64 - expected_size as f64) / expected_size as f64) * 100.0
             } else {
-                ((expected_size - actual_size) as f64 / expected_size as f64) * 100.0
+                ((expected_size as f64 - actual_size as f64) / expected_size as f64) * 100.0
             };
 
             // Allow 20% size difference to accommodate different versions
